requirements.txt
numpy
python-mpd2
requests
sounddevice
